Add import names for external deps in modules
We implemented external dependencies in modules to be only a list of
identifiers. In this case, the name under which a library is imported
in a module is the same as the name of the dependency that is announced
to the outside world.

But we may not want the two to collapse in all cases. With regular
Spaghetti modules, we ONLY need the module (package) identifier, as
modules are compiled so that in the host language each module may
import any other module under whatever name it desires.

Because for external dependencies we don't generate any proxies that can
be imported by language-native means, we are doing both package
provisioning (for this we need a package identifier), as well as an
automatic import (for this we need a variable identifier). It is
impractical if those two are forced to be the same.