Generate abstract types for transitive enums
Generate only abstract types for enums from transitive dependencies.

This causes enum generation to happen in three variants:
  - regular for locally defined enums
  - with indirection for enums defined in direct dependency
  - without member access (abstract type over integers) for enums
defined in transitive dependency

This seems like a lot of complexity, but I am coming to believe that it
is necessary complexity. It mirrors the three forms of module
generations for local, direct and transitive modules, where we generate
the full definition, then the accessor, then only the types (and right
now, erroneously I think, the enums).