Revise syntax of object types and method defs
This brings the language syntax closer to our main target languages and
improves the legibility of method definitions.

On the one hand, we want the syntax to follow the main conventions of
Haxe and TypeScript in ordering names and type annotations. On the
other hand, we don't align syntax too closely with either language in
order to provide signifiers that help prevent confusion about which
language is currently being edited.

The new function type syntax of `(A, B) -> R` avoids the confusion of
curried and multi-parameter functions.