Add three-tiered const generators
This makes const generation for Haxe and TypeScript work in the same way
as enum generation, in that const types from direct dependencies are
defined via reference to the dependency. Nothing is generated for
transitive dependencies, seeing that const "types" don't actually
declare a type that hasn't been priorly defined, and there are no values
to access.

The actual indirection is deactivated in this patch, to be activated
after a migratory period required to expose consts on module proxies.

Closes #206.